[Initilze Express app]
